Code cleanup for routing
