+++ src/routes/api.rs
@@ -7,8 +7,6 @@ use actix_web::{
     HttpRequest, HttpResponse, Responder,
 };
 
-use crate::utils::static_resolver::handle_embedded_file;
-
 use crate::{
     db::{
         get_picture_data_by_id, handle_filter_action, handle_get_all_action, handle_song_action,
@@ -19,7 +17,6 @@ use crate::{
 
 type DataDB = Data<r2d2::Pool<r2d2_sqlite::SqliteConnectionManager>>;
 
-// GET songs
 #[get("/songs")]
 pub async fn get_songs(db: DataDB) -> HttpResponse {
     let res = handle_song_action(&db, SongActions::GetAllSongs)
@@ -111,13 +108,14 @@ pub async fn get_picture(data: Path<String>, db: DataDB) -> HttpResponse {
     }
 }
 
-// View Routes
-#[get("/")]
-pub async fn index() -> HttpResponse {
-    handle_embedded_file("index.html")
-}
-
-#[get("/dist/{_:.*}")]
-async fn dist(path: Path<String>) -> impl Responder {
-    handle_embedded_file(path.as_str())
+pub fn register(config: &mut actix_web::web::ServiceConfig) {
+    config
+        .service(get_songs)
+        .service(get_song_by_id)
+        .service(get_albums)
+        .service(get_songs_by_album)
+        .service(get_artists)
+        .service(get_songs_by_artist)
+        .service(get_stream_by_id)
+        .service(get_picture);
 }
